refactor(cache): remove deprecated Cache class
The Cache class has been removed as its functionality is now handled by the Repository class.
This change helps streamline the codebase and removes duplicate implementations.
