+++ src/cache/src/CacheInterface.php
@@ -11,140 +11,13 @@ declare(strict_types=1);
 
 namespace FriendsOfHyperf\Cache;
 
-use Closure;
-use DateInterval;
-use DateTimeInterface;
+class_alias(Contract\Repository::class, CacheInterface::class);
 
-interface CacheInterface
-{
+if (false) { // @phpstan-ignore-line
     /**
-     * @param string $key
-     * @param mixed $value
-     * @param DateInterval|DateTimeInterface|int|null $ttl
+     * @deprecated since v3.1, use `\FriendsOfHyperf\Cache\Contract\Repository` instead, will removed in v3.2
      */
-    public function add($key, $value, $ttl = null): bool;
-
-    /**
-     * Retrieve an item from the cache by key, refreshing it in the background if it is stale.
-     *
-     * @template TCacheValue
-     *
-     * @param string $key
-     * @param array{ 0: DateTimeInterface|DateInterval|int, 1: DateTimeInterface|DateInterval|int } $ttl
-     * @param (callable(): TCacheValue) $callback
-     * @param array{ seconds?: int, owner?: string }|null $lock
-     * @return TCacheValue
-     */
-    public function flexible($key, $ttl, $callback, $lock = null);
-
-    public function flush(): bool;
-
-    /**
-     * @param string $key
-     * @param mixed $value
-     */
-    public function forever($key, $value): bool;
-
-    /**
-     * @param string $key
-     */
-    public function forget($key): bool;
-
-    /**
-     * Determine if an item exists in the cache.
-     *
-     * @param string $key
-     */
-    public function has($key): bool;
-
-    /**
-     * Determine if an item doesn't exist in the cache.
-     *
-     * @param string $key
-     */
-    public function missing($key): bool;
-
-    /**
-     * @param array|string $key
-     * @param mixed $value
-     * @param DateInterval|DateTimeInterface|int|null $ttl
-     */
-    public function put($key, $value, $ttl = null): bool;
-
-    public function putMany(array $values, $ttl = null): bool;
-
-    /**
-     * @param string $key
-     * @param int $value
-     * @return bool|int
-     */
-    public function decrement($key, $value = 1);
-
-    /**
-     * @param string $key
-     * @param int $value
-     * @return bool|int
-     */
-    public function increment($key, $value = 1);
-
-    /**
-     * Retrieve an item from the cache by key.
-     *
-     * @template TCacheValue
-     *
-     * @param string[]|string $key
-     * @param (Closure(): TCacheValue)|TCacheValue $default
-     * @return (TCacheValue is null ? mixed : TCacheValue)
-     */
-    public function get($key, $default = null);
-
-    /**
-     * @return iterable
-     */
-    public function many(array $keys);
-
-    /**
-     * Retrieve an item from the cache and delete it.
-     *
-     * @template TCacheValue
-     *
-     * @param string[]|string $key
-     * @param (Closure(): TCacheValue)|TCacheValue $default
-     * @return (TCacheValue is null ? mixed : TCacheValue)
-     */
-    public function pull($key, $default = null);
-
-    /**
-     * Get an item from the cache, or execute the given Closure and store the result.
-     *
-     * @template TCacheValue
-     *
-     * @param string $key
-     * @param DateInterval|DateTimeInterface|int|null $ttl
-     * @param Closure(): TCacheValue $callback
-     * @return TCacheValue
-     */
-    public function remember($key, $ttl, Closure $callback);
-
-    /**
-     * Get an item from the cache, or execute the given Closure and store the result forever.
-     *
-     * @template TCacheValue
-     *
-     * @param string $key
-     * @param Closure(): TCacheValue $callback
-     * @return TCacheValue
-     */
-    public function rememberForever($key, Closure $callback);
-
-    /**
-     * Get an item from the cache, or execute the given Closure and store the result forever.
-     *
-     * @template TCacheValue
-     *
-     * @param string $key
-     * @param Closure(): TCacheValue $callback
-     * @return TCacheValue
-     */
-    public function sear($key, Closure $callback);
+    interface CacheInterface extends Contract\Repository
+    {
+    }
 }
