+++ src/cache/src/Repository.php
@@ -37,7 +37,7 @@ use function Hyperf\Coroutine\defer;
 use function Hyperf\Support\with;
 use function Hyperf\Tappable\tap;
 
-class Cache implements CacheInterface
+class Repository implements Contract\Repository
 {
     use InteractsWithTime;
     use Macroable;
